+++ doc/version-history.rst
@@ -8,6 +8,7 @@ v1.0.1
 ------
 
 * ApplySetpoint command, auto-regulate the mixing valve
+* Auto-recover ILCs after bus failure
 * Improved error response on pump readout
 
 v1.0.0